+++ sharding-jdbc-orchestration/src/main/java/io/shardingjdbc/orchestration/api/OrchestrationShardingDataSourceFactory.java
@@ -20,104 +20,131 @@ package io.shardingjdbc.orchestration.api;
 import io.shardingjdbc.core.api.ShardingDataSourceFactory;
 import io.shardingjdbc.core.api.config.ShardingRuleConfiguration;
 import io.shardingjdbc.core.jdbc.core.datasource.ShardingDataSource;
-import io.shardingjdbc.orchestration.json.DataSourceJsonConverter;
-import io.shardingjdbc.orchestration.json.ShardingRuleConfigurationConverter;
-import io.shardingjdbc.orchestration.reg.base.CoordinatorRegistryCenter;
-import com.google.common.base.Charsets;
+import io.shardingjdbc.orchestration.api.config.OrchestrationConfiguration;
+import io.shardingjdbc.orchestration.internal.OrchestrationFacade;
+import io.shardingjdbc.orchestration.yaml.YamlOrchestrationShardingRuleConfiguration;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.curator.framework.CuratorFramework;
-import org.apache.curator.framework.recipes.cache.ChildData;
-import org.apache.curator.framework.recipes.cache.TreeCache;
-import org.apache.curator.framework.recipes.cache.TreeCacheEvent;
-import org.apache.curator.framework.recipes.cache.TreeCacheListener;
+import org.yaml.snakeyaml.Yaml;
+import org.yaml.snakeyaml.constructor.Constructor;
 
 import javax.sql.DataSource;
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStreamReader;
 import java.sql.SQLException;
 import java.util.Map;
 import java.util.Properties;
 
 /**
  * Orchestration sharding data source factory.
- * 
- * @author zhangliang 
+ *
+ * @author zhangliang
+ * @author caohao
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class OrchestrationShardingDataSourceFactory {
     
     /**
      * Create sharding data source.
-     * 
-     * @param name name of sharding data source
-     * @param registryCenter registry center
+     *
      * @param dataSourceMap data source map
-     * @param shardingRuleConfig rule configuration for databases and tables sharding
+     * @param shardingRuleConfig sharding rule configuration
+     * @param orchestrationConfig orchestration master-slave configuration
+     *
      * @return sharding data source
      * @throws SQLException SQL exception
      */
     public static DataSource createDataSource(
-            final String name, final CoordinatorRegistryCenter registryCenter, final Map<String, DataSource> dataSourceMap, final ShardingRuleConfiguration shardingRuleConfig) throws SQLException {
-        initRegistryCenter(name, registryCenter, dataSourceMap, shardingRuleConfig);
+            final Map<String, DataSource> dataSourceMap, final ShardingRuleConfiguration shardingRuleConfig, final OrchestrationConfiguration orchestrationConfig) throws SQLException {
+        return createDataSource(dataSourceMap, shardingRuleConfig, orchestrationConfig, new Properties());
+    }
+    
+    /**
+     * Create sharding data source.
+     *
+     * @param dataSourceMap data source map
+     * @param shardingRuleConfig sharding rule configuration
+     * @param orchestrationConfig orchestration master-slave configuration
+     * @param props properties for data source
+     *
+     * @return sharding data source
+     * @throws SQLException SQL exception
+     */
+    public static DataSource createDataSource(
+            final Map<String, DataSource> dataSourceMap, final ShardingRuleConfiguration shardingRuleConfig, 
+            final OrchestrationConfiguration orchestrationConfig, final Properties props) throws SQLException {
         ShardingDataSource result = (ShardingDataSource) ShardingDataSourceFactory.createDataSource(dataSourceMap, shardingRuleConfig);
-        addConfigurationChangeListener(name, registryCenter, result);
+        new OrchestrationFacade(orchestrationConfig).initShardingOrchestration(dataSourceMap, shardingRuleConfig, props, result);
         return result;
     }
     
     /**
      * Create sharding data source.
-     * 
-     * @param name name of sharding data source
-     * @param registryCenter registry center
-     * @param dataSourceMap data source map
-     * @param shardingRuleConfig rule configuration for databases and tables sharding
-     * @param props properties for data source
+     *
+     * @param yamlFile yaml file for rule configuration of databases and tables sharding with data sources
      * @return sharding data source
      * @throws SQLException SQL exception
+     * @throws IOException IO exception
      */
-    public static DataSource createDataSource(
-            final String name, final CoordinatorRegistryCenter registryCenter, final Map<String, DataSource> dataSourceMap,
-            final ShardingRuleConfiguration shardingRuleConfig, final Properties props) throws SQLException {
-        initRegistryCenter(name, registryCenter, dataSourceMap, shardingRuleConfig);
-        // TODO props
-        ShardingDataSource result = (ShardingDataSource) ShardingDataSourceFactory.createDataSource(dataSourceMap, shardingRuleConfig, props);
-        addConfigurationChangeListener(name, registryCenter, result);
-        return result;
+    public static DataSource createDataSource(final File yamlFile) throws SQLException, IOException {
+        YamlOrchestrationShardingRuleConfiguration config = unmarshal(yamlFile);
+        return createDataSource(config.getDataSources(), config.getShardingRule().getShardingRuleConfiguration(), config.getOrchestration().getOrchestrationConfiguration());
     }
     
-    private static void initRegistryCenter(final String name, 
-                                           final CoordinatorRegistryCenter registryCenter, final Map<String, DataSource> dataSourceMap, final ShardingRuleConfiguration shardingRuleConfig) {
-        registryCenter.init();
-        registryCenter.persist("/" + name + "/config/datasource", DataSourceJsonConverter.toJson(dataSourceMap));
-        registryCenter.persist("/" + name + "/config/sharding", ShardingRuleConfigurationConverter.toJson(shardingRuleConfig));
-        registryCenter.addCacheData("/" + name + "/config");
+    /**
+     * Create sharding data source.
+     *
+     * @param dataSourceMap data source map
+     * @param yamlFile yaml file for rule configuration of databases and tables sharding without data sources
+     * @return sharding data source
+     * @throws SQLException SQL exception
+     * @throws IOException IO exception
+     */
+    public static DataSource createDataSource(final Map<String, DataSource> dataSourceMap, final File yamlFile) throws SQLException, IOException {
+        YamlOrchestrationShardingRuleConfiguration config = unmarshal(yamlFile);
+        return createDataSource(dataSourceMap,  config.getShardingRule().getShardingRuleConfiguration(), config.getOrchestration().getOrchestrationConfiguration());
     }
     
-    private static void addConfigurationChangeListener(final String name, final CoordinatorRegistryCenter registryCenter, final ShardingDataSource shardingDataSource) {
-        TreeCache cache = (TreeCache) registryCenter.getRawCache("/" + name + "/config");
-        cache.getListenable().addListener(new TreeCacheListener() {
-            
-            @Override
-            public void childEvent(final CuratorFramework client, final TreeCacheEvent event) throws Exception {
-                ChildData childData = event.getData();
-                if (null == childData || null == childData.getData()) {
-                    return;
-                }
-                String path = childData.getPath();
-                if (path.isEmpty()) {
-                    return;
-                }
-                if (("/" + name + "/config/datasource").equals(path)) {
-                    Map<String, DataSource> newDataSourceMap = DataSourceJsonConverter.fromJson(new String(childData.getData(), Charsets.UTF_8));
-                    ShardingRuleConfiguration shardingRuleConfig = ShardingRuleConfigurationConverter.fromJson(registryCenter.get("/" + name + "/config/sharding"));
-                    // TODO props
-                    shardingDataSource.renew(shardingRuleConfig.build(newDataSourceMap), new Properties());
-                } else if (("/" + name + "/config/sharding").equals(path)) {
-                    ShardingRuleConfiguration newShardingRuleConfig = ShardingRuleConfigurationConverter.fromJson(new String(childData.getData(), Charsets.UTF_8));
-                    Map<String, DataSource> dataSourceMap = DataSourceJsonConverter.fromJson(registryCenter.get("/" + name + "/config/datasource"));
-                    // TODO props
-                    shardingDataSource.renew(newShardingRuleConfig.build(dataSourceMap), new Properties());
-                }
-            }
-        });
+    /**
+     * Create sharding data source.
+     *
+     * @param yamlByteArray yaml byte array for rule configuration of databases and tables sharding with data sources
+     * @return sharding data source
+     * @throws SQLException SQL exception
+     * @throws IOException IO exception
+     */
+    public static DataSource createDataSource(final byte[] yamlByteArray) throws SQLException, IOException {
+        YamlOrchestrationShardingRuleConfiguration config = unmarshal(yamlByteArray);
+        return createDataSource(config.getDataSources(),  config.getShardingRule().getShardingRuleConfiguration(), config.getOrchestration().getOrchestrationConfiguration());
+    }
+    
+    /**
+     * Create sharding data source.
+     *
+     * @param dataSourceMap data source map
+     * @param yamlByteArray yaml byte array for rule configuration of databases and tables sharding without data sources
+     * @return sharding data source
+     * @throws SQLException SQL exception
+     * @throws IOException IO exception
+     */
+    public static DataSource createDataSource(final Map<String, DataSource> dataSourceMap, final byte[] yamlByteArray) throws SQLException, IOException {
+        YamlOrchestrationShardingRuleConfiguration config = unmarshal(yamlByteArray);
+        return createDataSource(dataSourceMap,  config.getShardingRule().getShardingRuleConfiguration(), config.getOrchestration().getOrchestrationConfiguration());
+    }
+    
+    private static YamlOrchestrationShardingRuleConfiguration unmarshal(final File yamlFile) throws IOException {
+        try (
+                FileInputStream fileInputStream = new FileInputStream(yamlFile);
+                InputStreamReader inputStreamReader = new InputStreamReader(fileInputStream, "UTF-8")
+        ) {
+            return new Yaml(new Constructor(YamlOrchestrationShardingRuleConfiguration.class)).loadAs(inputStreamReader, YamlOrchestrationShardingRuleConfiguration.class);
+        }
+    }
+    
+    private static YamlOrchestrationShardingRuleConfiguration unmarshal(final byte[] yamlByteArray) throws IOException {
+        return new Yaml(new Constructor(YamlOrchestrationShardingRuleConfiguration.class)).loadAs(new ByteArrayInputStream(yamlByteArray), YamlOrchestrationShardingRuleConfiguration.class);
     }
 }