+++ sharding-jdbc-spring/sharding-jdbc-spring-namespace/src/test/java/io/shardingjdbc/spring/ShardingNamespaceTest.java
@@ -17,10 +17,6 @@
 
 package io.shardingjdbc.spring;
 
-import org.junit.Test;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.AbstractJUnit4SpringContextTests;
-
 import io.shardingjdbc.core.api.config.strategy.ComplexShardingStrategyConfiguration;
 import io.shardingjdbc.core.api.config.strategy.HintShardingStrategyConfiguration;
 import io.shardingjdbc.core.api.config.strategy.InlineShardingStrategyConfiguration;
@@ -35,6 +31,12 @@ import io.shardingjdbc.core.rule.ShardingRule;
 import io.shardingjdbc.core.rule.TableRule;
 import io.shardingjdbc.spring.datasource.SpringShardingDataSource;
 import io.shardingjdbc.spring.util.FieldValueUtil;
+import org.junit.Test;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.AbstractJUnit4SpringContextTests;
+
+import java.util.Arrays;
+import java.util.Iterator;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertNotNull;
@@ -42,21 +44,18 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
-import java.util.Arrays;
-import java.util.Iterator;
-
 @ContextConfiguration(locations = "classpath:META-INF/rdb/shardingNamespace.xml")
 public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     
     @Test
-    public void testStandardStrategy() {
+    public void assertStandardStrategy() {
         StandardShardingStrategyConfiguration standardStrategy = this.applicationContext.getBean("standardStrategy", StandardShardingStrategyConfiguration.class);
         assertThat(standardStrategy.getShardingColumn(), is("user_id"));
         assertThat(standardStrategy.getPreciseAlgorithmClassName(), is("io.shardingjdbc.spring.algorithm.PreciseModuloDatabaseShardingAlgorithm"));
     }
     
     @Test
-    public void testRangeStandardStrategy() {
+    public void assertRangeStandardStrategy() {
         StandardShardingStrategyConfiguration rangeStandardStrategy = this.applicationContext.getBean("rangeStandardStrategy", StandardShardingStrategyConfiguration.class);
         assertThat(rangeStandardStrategy.getShardingColumn(), is("order_id"));
         assertThat(rangeStandardStrategy.getPreciseAlgorithmClassName(), is("io.shardingjdbc.spring.algorithm.PreciseModuloTableShardingAlgorithm"));
@@ -64,32 +63,32 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testComplexStrategy() {
+    public void assertComplexStrategy() {
         ComplexShardingStrategyConfiguration complexStrategy = this.applicationContext.getBean("complexStrategy", ComplexShardingStrategyConfiguration.class);
         assertThat(complexStrategy.getShardingColumns(), is("order_id,user_id"));
         assertThat(complexStrategy.getAlgorithmClassName(), is("io.shardingjdbc.spring.algorithm.DefaultComplexKeysShardingAlgorithm"));
     }
     
     @Test
-    public void testInlineStrategy() {
+    public void assertInlineStrategy() {
         InlineShardingStrategyConfiguration inlineStrategy = this.applicationContext.getBean("inlineStrategy", InlineShardingStrategyConfiguration.class);
         assertThat(inlineStrategy.getShardingColumn(), is("order_id"));
-        assertThat(inlineStrategy.getAlgorithmInlineExpression(), is("t_order_${order_id % 4}"));
+        assertThat(inlineStrategy.getAlgorithmExpression(), is("t_order_${order_id % 4}"));
     }
     
     @Test
-    public void testHintStrategy() {
+    public void assertHintStrategy() {
         HintShardingStrategyConfiguration hintStrategy = this.applicationContext.getBean("hintStrategy", HintShardingStrategyConfiguration.class);
         assertThat(hintStrategy.getAlgorithmClassName(), is("io.shardingjdbc.spring.algorithm.DefaultHintShardingAlgorithm"));
     }
     
     @Test
-    public void testNoneStrategy() {
+    public void assertNoneStrategy() {
         this.applicationContext.getBean("noneStrategy", NoneShardingStrategyConfiguration.class);
     }
     
     @Test
-    public void testSimpleShardingDataSource() {
+    public void assertSimpleShardingDataSource() {
         ShardingRule shardingRule = getShardingRule("simpleShardingDataSource");
         assertNotNull(shardingRule.getDataSourceMap().get("dbtbl_0"));
         assertThat(shardingRule.getTableRules().size(), is(1));
@@ -97,7 +96,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testShardingRuleWithAttributesDataSource() {
+    public void assertShardingRuleWithAttributesDataSource() {
         ShardingRule shardingRule = getShardingRule("shardingRuleWithAttributesDataSource");
         assertNotNull(shardingRule.getDataSourceMap().get("dbtbl_0"));
         assertNotNull(shardingRule.getDataSourceMap().get("dbtbl_1"));
@@ -110,7 +109,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testTableRuleWithAttributesDataSource() {
+    public void assertTableRuleWithAttributesDataSource() {
         ShardingRule shardingRule = getShardingRule("tableRuleWithAttributesDataSource");
         assertThat(shardingRule.getTableRules().size(), is(1));
         TableRule tableRule = shardingRule.getTableRules().iterator().next();
@@ -133,7 +132,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testMultiTableRulesDataSource() {
+    public void assertMultiTableRulesDataSource() {
         ShardingRule shardingRule = getShardingRule("multiTableRulesDataSource");
         assertThat(shardingRule.getTableRules().size(), is(2));
         Iterator<TableRule> iter = shardingRule.getTableRules().iterator();
@@ -142,7 +141,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testBindingTableRuleDatasource() {
+    public void assertBindingTableRuleDatasource() {
         ShardingRule shardingRule = getShardingRule("bindingTableRuleDatasource");
         assertThat(shardingRule.getBindingTableRules().size(), is(1));
         BindingTableRule bindingTableRule = shardingRule.getBindingTableRules().iterator().next();
@@ -151,7 +150,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testMultiBindingTableRulesDatasource() {
+    public void assertMultiBindingTableRulesDatasource() {
         ShardingRule shardingRule = getShardingRule("multiBindingTableRulesDatasource");
         assertThat(shardingRule.getBindingTableRules().size(), is(2));
         Iterator<BindingTableRule> iter = shardingRule.getBindingTableRules().iterator();
@@ -164,7 +163,7 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testPropsDataSource() {
+    public void assertPropsDataSource() {
         ShardingDataSource shardingDataSource = this.applicationContext.getBean("propsDataSource", ShardingDataSource.class);
         Object shardingContext = FieldValueUtil.getFieldValue(shardingDataSource, "shardingContext", true);
         assertTrue((boolean) FieldValueUtil.getFieldValue(shardingContext, "showSQL"));
@@ -177,12 +176,12 @@ public class ShardingNamespaceTest extends AbstractJUnit4SpringContextTests {
     }
     
     @Test
-    public void testShardingDataSourceType() {
+    public void assertShardingDataSourceType() {
         assertTrue(this.applicationContext.getBean("simpleShardingDataSource", ShardingDataSource.class) instanceof SpringShardingDataSource);
     }
     
     @Test
-    public void testDefaultActualDataNodes() {
+    public void assertDefaultActualDataNodes() {
         ShardingDataSource multiTableRulesDataSource = this.applicationContext.getBean("multiTableRulesDataSource", ShardingDataSource.class);
         Object shardingContext = FieldValueUtil.getFieldValue(multiTableRulesDataSource, "shardingContext", true);
         ShardingRule shardingRule = (ShardingRule) FieldValueUtil.getFieldValue(shardingContext, "shardingRule");